Added Hasher class
Hasher class to centralize the hashing helper methods
